Refactor composer install to multi stage
When installing composer some security checks **SHOULD** be performed to make sure the correct, unmodified, version of composer is installed. This can be done manually, or we can use the prebuilt composer docker image which already takes care of this. If this all succeeds the installed composer phar file is being copied in the final build.